add login

+++ src/main/java/jpaboard/board/controller/BoardController.java
@@ -2,16 +2,16 @@ package jpaboard.board.controller;
 
 
 import jpaboard.board.domain.Board;
+import jpaboard.board.domain.Member;
 import jpaboard.board.service.BoardService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
+@SessionAttributes("member")
 @Controller
 public class BoardController {
 
@@ -23,8 +23,16 @@ public class BoardController {
         model.addAttribute("greeting","Hello 타임리프^^");
     }
 
+    @ModelAttribute("member")
+    public Member setMember(){
+        return new Member();
+    }
+
     @RequestMapping("/getBoardList")
-    public String getBoardList(Model model) {
+    public String getBoardList(@ModelAttribute("member") Member member, Model model) {
+        if(member.getId() == null){
+            return "redirect:login";
+        }
         List<Board> boardList = boardService.getBoardList();
 
         model.addAttribute("boardList", boardList);
@@ -32,31 +40,47 @@ public class BoardController {
     }
 
     @GetMapping("/insertBoard")
-    public String insertBoardView(){
+    public String insertBoardView(@ModelAttribute("member") Member member){
+        if(member.getId() == null){
+            return "redirect:login";
+        }
         return "insertBoard";
     }
 
     @PostMapping("/insertBoard")
-    public String insertBoard(Board board){
+    public String insertBoard(Board board, @ModelAttribute("member") Member member){
+        if(member.getId() == null){
+            return "redirect:login";
+        }
         boardService.insertBoard(board);
         return "redirect:getBoardList";
     }
 
     @GetMapping("/getBoard")
-    public String getBoard(Board board, Model model){
+    public String getBoard(Board board, Model model, @ModelAttribute("member") Member member){
+        if(member.getId() == null){
+            return "redirect:login";
+        }
         model.addAttribute("board",boardService.getBoard(board));
         return "getBoard";
     }
 
     @PostMapping("/updateBoard")
-    public String updateBoard(Board board){
+    public String updateBoard(Board board, @ModelAttribute("member") Member member){
+        if(member.getId() == null){
+            return "redirect:login";
+        }
         boardService.updateBoard(board);
         return "forward:getBoardList";
     }
 
     @GetMapping("/deleteBoard")
-    public String deleteBoard(Board board){
+    public String deleteBoard(Board board, @ModelAttribute("member") Member member){
+        if(member.getId() == null){
+            return "redirect:login";
+        }
         boardService.deleteBoard(board);
         return "forward:getBoardList";
     }
+
 }
